requirements.txt
langchain
langchain-core
langchain-community
fastapi
uvicorn
typing
python-dotenv
langchain-google-genai
langchain-pinecone
pymupdf4llm
PyMuPDF
pinecone-client=
sentence-transformers
python-multipart
pydantic
pillow
streamlit
requests
pandas